SAK-33981 refactor chart to wicket component (#5372)
* Remove unused code and cleanup other compiler warnings

* Add rebel.xml to gitignore so that JRebel project config is not committed

* SAK-33981 refactor chart to component so it can be reused.

(cherry picked from commit d3367ecba36c2164479c43f3f286bdf19cad4140)
